requirements.txt
tensorflow
numpy
flask
Pillow
Image
requests